🚑 Hotfix: Initialize every modal state
