+++ src/features/stock-ordering/presentation/components/index.tsx
@@ -9,3 +9,5 @@ export { StockOrderViewContent } from "./stock-order-view-content";
 export { OrderPlaceAndConfirmTable } from "./order-place-and-confirm-table";
 export { StockOrderTable } from "./stock-order-table";
 export { StockOrderLogs } from "./stock-order-logs";
+export { InitializeModal } from "./initialize-modal";
+export { InitializeProductData } from "./initialize-product-data";
