+++ src/features/stock-ordering/presentation/modals/store-receive-order.modal.tsx
@@ -5,18 +5,16 @@ import { useAppDispatch, useAppSelector } from "features/config/hooks";
 import { LocalizationProvider, DatePicker } from "@mui/x-date-pickers";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import dayjs from "dayjs";
-import { StockOrderTable } from "../components";
+import {
+  InitializeModal,
+  InitializeProductData,
+  StockOrderTable,
+} from "../components";
 import { UploadDeliveryRecieptModal } from "./upload-delivery-reciepts.modal";
 import { MdPreview } from "react-icons/md";
 import { TableRow } from "features/stock-ordering/core/domain/table-row.model";
-import {
-  getProductData,
-  selectGetProductData,
-} from "../slices/get-product-data.slice";
-import {
-  orderID,
-  receiveOrdersParam,
-} from "features/stock-ordering/core/stock-ordering.params";
+import { selectGetProductData } from "../slices/get-product-data.slice";
+import { receiveOrdersParam } from "features/stock-ordering/core/stock-ordering.params";
 import { updateReceiveOrders } from "../slices/update-receive-order.slice";
 
 interface StoreReceiveOrderModalProps {
@@ -34,7 +32,6 @@ export function StoreReceiveOrderModal(props: StoreReceiveOrderModalProps) {
   const [actualDeliveryDate, setActualDeliveryDate] = useState(
     dayjs().format("YYYY-MM-DD HH:mm:ss")
   );
-  const [isDeliveredQtyAvailable, setIsDeliveredQtyAvailable] = useState(true);
   const [openUploadDeliveryRecieptModal, setOpenUploadDeliveryRecieptModal] =
     useState(false);
 
@@ -60,56 +57,74 @@ export function StoreReceiveOrderModal(props: StoreReceiveOrderModalProps) {
     product_data: [],
   });
 
+  // useEffect(() => {
+  //   const productID: orderID = { orderId: props.id };
+
+  //   if (props.open) {
+  //     dispatch(getProductData(productID));
+  //   }
+  // }, [dispatch, props.open]);
+
+  // useEffect(() => {
+  //   if (
+  //     getProductDataState.data &&
+  //     getProductDataState.data.product_data &&
+  //     getProductDataState.data.order_information
+  //   ) {
+  //     const order = getProductDataState.data?.order_information;
+  //     const order_information: TableRow["order_information"] = {
+  //       store_name: order.store_name,
+  //       order_number: order.id,
+  //       requested_delivery_date: order.requested_delivery_date,
+  //       commited_delivery_date: order.commited_delivery_date,
+  //       order_reviewed_date: order.reviewed_date,
+  //       order_confirmation_date: order.order_confirmation_date,
+  //       view_delivery_receipt: order.delivery_receipt,
+  //       dispatch_date: order.dispatch_date,
+  //       order_enroute: order.enroute_date,
+  //       actual_delivery_date: order.actual_delivery_date,
+  //       view_updated_delivery_receipt: order.updated_delivery_receipt,
+  //       billing_information_ready:
+  //         order.billing_id && order.billing_amount ? true : false,
+  //       view_payment_details: order.payment_detail_image,
+  //       payment_confirmation: order.payment_confirmation_date,
+  //     };
+
+  //     const productData: TableRow["product_data"] =
+  //       getProductDataState.data.product_data.map((product) => ({
+  //         id: product.id,
+  //         productId: product.product_id,
+  //         productName: product.product_name,
+  //         uom: product.uom,
+  //         orderQty: product.order_qty,
+  //         commitedQuantity: product.commited_qty,
+  //         deliveredQuantity: product.delivered_qty,
+  //       }));
+
+  //     setRows({
+  //       order_information: order_information,
+  //       product_data: productData,
+  //     });
+  //   }
+  // }, [getProductDataState.data]);
+
   useEffect(() => {
-    const productID: orderID = { orderId: props.id };
+    setActualDeliveryDate(dayjs().format("YYYY-MM-DD HH:mm:ss"));
+    setUploadedReciept("");
+  }, [props.open]);
 
-    if (props.open) {
-      dispatch(getProductData(productID));
-    }
-  }, [dispatch, props.open]);
+  InitializeModal({
+    setRows: setRows,
+    id: props.id,
+    open: props.open,
+  });
 
-  useEffect(() => {
-    if (
-      getProductDataState.data &&
-      getProductDataState.data.product_data &&
-      getProductDataState.data.order_information
-    ) {
-      const order = getProductDataState.data?.order_information;
-      const order_information: TableRow["order_information"] = {
-        store_name: order.store_name,
-        order_number: order.id,
-        requested_delivery_date: order.requested_delivery_date,
-        commited_delivery_date: order.commited_delivery_date,
-        order_reviewed_date: order.reviewed_date,
-        order_confirmation_date: order.order_confirmation_date,
-        view_delivery_receipt: order.delivery_receipt,
-        dispatch_date: order.dispatch_date,
-        order_enroute: order.enroute_date,
-        actual_delivery_date: order.actual_delivery_date,
-        view_updated_delivery_receipt: order.updated_delivery_receipt,
-        billing_information_ready:
-          order.billing_id && order.billing_amount ? true : false,
-        view_payment_details: order.payment_detail_image,
-        payment_confirmation: order.payment_confirmation_date,
-      };
-
-      const productData: TableRow["product_data"] =
-        getProductDataState.data.product_data.map((product) => ({
-          id: product.id,
-          productId: product.product_id,
-          productName: product.product_name,
-          uom: product.uom,
-          orderQty: product.order_qty,
-          commitedQuantity: product.commited_qty,
-          deliveredQuantity: product.delivered_qty,
-        }));
-
-      setRows({
-        order_information: order_information,
-        product_data: productData,
-      });
-    }
-  }, [getProductDataState.data]);
+  InitializeProductData({
+    setRows: setRows,
+    productData: getProductDataState.data
+      ? getProductDataState.data
+      : undefined,
+  });
 
   const handleSubmit = async (e: { preventDefault: () => void }) => {
     e.preventDefault();
@@ -131,8 +146,8 @@ export function StoreReceiveOrderModal(props: StoreReceiveOrderModalProps) {
 
       await dispatch(updateReceiveOrders(receiveOrdersParamData));
 
-      setIsDeliveredQtyAvailable(false);
       setButtonDisable(true);
+      props.onClose();
     } else {
       setOpenUploadDeliveryRecieptModal(true);
     }
@@ -171,8 +186,6 @@ export function StoreReceiveOrderModal(props: StoreReceiveOrderModalProps) {
     return null;
   }
 
-  console.log(uploadedReceipt);
-
   return (
     <>
       <div className="fixed inset-0 z-30 flex items-start justify-center overflow-auto bg-black bg-opacity-30 backdrop-blur-sm">
@@ -198,7 +211,7 @@ export function StoreReceiveOrderModal(props: StoreReceiveOrderModalProps) {
                 activeTab={props.currentTab}
                 setRows={setRows}
                 rowData={rows}
-                isDeliveredQtyAvailable={isDeliveredQtyAvailable}
+                isDeliveredQtyAvailable={true}
               />
 
               {buttonDisable ? null : (
