+++ src/features/stock-ordering/presentation/modals/supplier-en-route-order.modal.tsx
@@ -17,6 +17,7 @@ import {
   updateStatus,
 } from "features/stock-ordering/core/stock-ordering.params";
 import { updateEnrouteOrders } from "../slices/update-enroute-order.slice";
+import { InitializeModal, InitializeProductData } from "../components";
 
 interface SupplierEnRouteOrderModalProps {
   open: boolean;
@@ -55,56 +56,69 @@ export function SupplierEnRouteOrderModal(
     product_data: [],
   });
 
-  useEffect(() => {
-    const productID: orderID = { orderId: props.id };
+  // useEffect(() => {
+  //   const productID: orderID = { orderId: props.id };
 
-    if (props.open) {
-      dispatch(getProductData(productID));
-    }
-  }, [dispatch, props.open]);
+  //   if (props.open) {
+  //     dispatch(getProductData(productID));
+  //   }
+  // }, [dispatch, props.open]);
 
-  useEffect(() => {
-    if (
-      getProductDataState.data &&
-      getProductDataState.data.product_data &&
-      getProductDataState.data.order_information
-    ) {
-      const order = getProductDataState.data?.order_information;
-      const order_information: TableRow["order_information"] = {
-        store_name: order.store_name,
-        order_number: order.id,
-        requested_delivery_date: order.requested_delivery_date,
-        commited_delivery_date: order.commited_delivery_date,
-        order_reviewed_date: order.reviewed_date,
-        order_confirmation_date: order.order_confirmation_date,
-        view_delivery_receipt: order.delivery_receipt,
-        dispatch_date: order.dispatch_date,
-        order_enroute: order.enroute_date,
-        actual_delivery_date: order.actual_delivery_date,
-        view_updated_delivery_receipt: order.updated_delivery_receipt,
-        billing_information_ready:
-          order.billing_id && order.billing_amount ? true : false,
-        view_payment_details: order.payment_detail_image,
-        payment_confirmation: order.payment_confirmation_date,
-      };
+  // useEffect(() => {
+  //   if (
+  //     getProductDataState.data &&
+  //     getProductDataState.data.product_data &&
+  //     getProductDataState.data.order_information
+  //   ) {
+  //     const order = getProductDataState.data?.order_information;
+  //     const order_information: TableRow["order_information"] = {
+  //       store_name: order.store_name,
+  //       order_number: order.id,
+  //       requested_delivery_date: order.requested_delivery_date,
+  //       commited_delivery_date: order.commited_delivery_date,
+  //       order_reviewed_date: order.reviewed_date,
+  //       order_confirmation_date: order.order_confirmation_date,
+  //       view_delivery_receipt: order.delivery_receipt,
+  //       dispatch_date: order.dispatch_date,
+  //       order_enroute: order.enroute_date,
+  //       actual_delivery_date: order.actual_delivery_date,
+  //       view_updated_delivery_receipt: order.updated_delivery_receipt,
+  //       billing_information_ready:
+  //         order.billing_id && order.billing_amount ? true : false,
+  //       view_payment_details: order.payment_detail_image,
+  //       payment_confirmation: order.payment_confirmation_date,
+  //     };
 
-      const productData: TableRow["product_data"] =
-        getProductDataState.data.product_data.map((product) => ({
-          id: product.id,
-          productId: product.product_id,
-          productName: product.product_name,
-          uom: product.uom,
-          orderQty: product.order_qty,
-          commitedQuantity: product.commited_qty,
-          deliveredQuantity: product.delivered_qty,
-        }));
+  //     const productData: TableRow["product_data"] =
+  //       getProductDataState.data.product_data.map((product) => ({
+  //         id: product.id,
+  //         productId: product.product_id,
+  //         productName: product.product_name,
+  //         uom: product.uom,
+  //         orderQty: product.order_qty,
+  //         commitedQuantity: product.commited_qty,
+  //         deliveredQuantity: product.delivered_qty,
+  //       }));
 
-      setRows({
-        order_information: order_information,
-        product_data: productData,
-      });
-    }
-  }, [getProductDataState.data]);
+  //     setRows({
+  //       order_information: order_information,
+  //       product_data: productData,
+  //     });
+  //   }
+  // }, [getProductDataState.data]);
+
+  InitializeModal({
+    setRows: setRows,
+    id: props.id,
+    open: props.open,
+  });
+
+  InitializeProductData({
+    setRows: setRows,
+    productData: getProductDataState.data
+      ? getProductDataState.data
+      : undefined,
+  });
 
   const handleEnRoute = async () => {
     const enRouteOrdersParamData: updateStatus = {
